Added tests for cell filtering.
